requirements.txt
langchain
langchain_core
langchain_community
langchain_ollama
langchain-text-splitters
pypdf
faiss-cpu
python-dotenv
streamlit